Undo merge of coordinator and hub.
