+++ custom_components/button_plus/__init__.py
@@ -6,10 +6,10 @@ import logging
 from homeassistant.config_entries import ConfigEntry
 from homeassistant.core import HomeAssistant
 
-from . import buttonplushub
 from .button_plus_api.model import DeviceConfiguration
-from .buttonplushub import ButtonPlusHub, ButtonPlusCoordinator
+from .buttonplushub import ButtonPlusHub
 from .const import DOMAIN
+from .coordinator import ButtonPlusCoordinator
 
 _LOGGER = logging.getLogger(__name__)
