+++ custom_components/button_plus/buttonplushub.py
@@ -2,13 +2,9 @@
 from __future__ import annotations
 
 import logging
-import re
 
-from homeassistant.components.button import ButtonEntity
-from homeassistant.components.mqtt import ReceiveMessage, client as mqtt
 from homeassistant.config_entries import ConfigEntry
 from homeassistant.helpers import device_registry as dr
-from homeassistant.helpers.update_coordinator import DataUpdateCoordinator
 
 from .button_plus_api.local_api_client import LocalApiClient
 from .button_plus_api.model import DeviceConfiguration
@@ -65,48 +61,3 @@ class ButtonPlusHub:
 
     def add_top_label(self, button_id, entity):
         self.top_label_entities[str(button_id)] = entity
-
-
-class ButtonPlusCoordinator(DataUpdateCoordinator):
-    """Button Plus coordinator."""
-
-    def __init__(self, hass: HomeAssistant, hub: ButtonPlusHub):
-        """Initialize my coordinator."""
-        super().__init__(
-            hass,
-            _LOGGER,
-            name=f"{DOMAIN}_coordinator",
-            update_interval=None,
-            update_method=None,
-        )
-        self._hass = hass
-        self.hub = hub
-        self._mqtt_subscribed_buttons = False
-        self._mqtt_topic_buttons = "buttonplus/+/button/+/click"
-
-    async def _async_update_data(self):
-        """Create MQTT subscriptions for buttonplus """
-        _LOGGER.debug(f"Initial data fetch from coordinator")
-        if not self._mqtt_subscribed_buttons:
-            self.unsubscribe_mqtt = await mqtt.async_subscribe(
-                self._hass,
-                self._mqtt_topic_buttons,
-                self.mqtt_button_callback,
-                0
-            )
-            _LOGGER.debug(f"MQTT subscribed to {self._mqtt_topic_buttons}")
-
-    async def mqtt_button_callback(self, message: ReceiveMessage):
-        # Handle the message here
-        _LOGGER.debug(f"Received message on topic {message.topic}: {message.payload}")
-        match = re.search(r'/(\d+)/click', message.topic)
-        btn_id = int(match.group(1)) if match else None
-
-        entity: ButtonEntity = self.hub.button_entities[str(btn_id)]
-
-        await self.hass.services.async_call(
-            "button",
-            'press',
-            target={"entity_id": entity.entity_id}
-        )
-
